feat: food vending system

+++ src/main/java/br/hallel/relational/api/app/event/dto/EventFoodSaleDTO.java
@@ -1,10 +1,7 @@
 package br.hallel.relational.api.app.event.dto;
 
-import java.math.BigDecimal;
-import java.util.UUID;
-
 public record EventFoodSaleDTO(
          Integer quantity,
-         BigDecimal price
+         Double price
 ) {
 }
