+++ src/main/java/br/hallel/relational/api/app/event/dto/FoodRequestDTO.java
@@ -5,7 +5,6 @@ import jakarta.validation.constraints.DecimalMin;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -15,7 +14,7 @@ public record FoodRequestDTO(
         String name,
 
         @NotNull @DecimalMin(value = "0.01", message = "Total amount must be positive")
-        BigDecimal value,
+        Double value,
 
         @NotNull @DecimalMin(value = "1", message = "Total amount must be positive")
         Integer stockQuantity,
